IOW-307 Fixing few peer review comments from Mary.
