+++ assets/src/scripts/components/hydrograph/layout.js
@@ -8,8 +8,6 @@ import config from '../../config';
 import { mediaQuery } from '../../utils';
 import { tickSelector } from './domain';
 
-
-
 export const ASPECT_RATIO = 1 / 2;
 export const ASPECT_RATIO_PERCENT = `${100 * ASPECT_RATIO}%`;
 const MARGIN = {
@@ -24,7 +22,7 @@ const MARGIN_SMALL_DEVICE = {
     bottom: 10,
     left: 0
 };
-export const CIRCLE_RADIUS = 4;
+//export const CIRCLE_RADIUS = 4;
 export const CIRCLE_RADIUS_SINGLE_PT = 1;
 
 export const BRUSH_HEIGHT = 100;
